Move tutorial tools to message_ix.util.tutorial
