--- tutorial/Austrian_energy_system/tools.py
@@ -1,7 +1,0 @@
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parents[1] / "utils"))
-
-from plotting import Plots  # noqa: E402, F401
-from run_scenarios import *  # noqa: E402, F401, F403
