--- tutorial/westeros/tools.py
@@ -1,8 +1,0 @@
-import sys
-from pathlib import Path
-
-sys.path.append(str(Path(__file__).parents[1] / "utils"))
-
-from plotting import prepare_tutorial_plots  # noqa: E402
-
-__all__ = ["prepare_tutorial_plots"]
